Refactor e-board code
All e-board code except for DGT is moved to the eboard package. Code duplication is reduced as well.
