+++ eboard/chessnut/protocol.py
@@ -13,32 +13,28 @@
 #
 # Some code copied from https://github.com/domschl/python-mchess/blob/master/mchess/chess_link.py
 
-from typing import List
 
-import json
-import logging
-import queue
 import time
+import logging
 import threading
-import typing
+import queue
+import json
 
-from move_debouncer import MoveDebouncer
-from certabo import command
-from certabo.parser import CalibrationCallback, CertaboCalibrator, CertaboPiece, \
-    CertaboBoardMessageParser, ParserCallback
-from certabo.led_control import CertaboLedControl
-from certabo.sentio import Sentio
-from certabo.usb_transport import Transport
-
+from eboard.move_debouncer import MoveDebouncer
+from eboard.ble_transport import Transport
+from eboard.chessnut.parser import Parser, ParserCallback, Battery
+from eboard.chessnut import command
 
 logger = logging.getLogger(__name__)
+READ_CHARACTERISTIC = '1b7e8262-2877-41c3-b46e-cf057c562023'
+WRITE_CHARACTERISTIC = '1b7e8272-2877-41c3-b46e-cf057c562023'
 
 
-class Protocol(ParserCallback, CalibrationCallback):
+class Protocol(ParserCallback):
     """
-    This implements the 'Certabo' protocol for Certabo e-boards.
+    This implements the 'Chessnut' protocol for Chessnut e-boards.
 
-    Communication with the board is asynchronous. Replies from the board are written
+    Communcation with the board is asynchronous. Replies from the board are written
     to the python queue (`appqueue`) that is provided during instantiation.
 
     Every message in `appqueue` is a short json string.
@@ -50,25 +46,22 @@ class Protocol(ParserCallback, CalibrationCallback):
         :param name: identifies this protocol
         """
         super().__init__()
-        self.piece_recognition = False
         self.name = name
-        logger.debug('Certabo starting')
+        logger.debug('Chessnut starting')
         self.error_condition = False
         self.appque = appque
         self.board_mutex = threading.Lock()
         self.trque = queue.Queue()
         self.trans = None
-        self.led_control = None
+        self.config = None
         self.connected = False
         self.last_fen = None
-        self.low_gain = True
-        self.calibrator = CertaboCalibrator(self)
-        self.calibrated = False
-        self.initial_position_received = False  # initial position after calibration is complete
+        self.parser = Parser(self)
         self.brd_reversed = False
         self.device_in_config = False
         self.debouncer = MoveDebouncer(350, lambda fen: self.appque.put({'cmd': 'raw_board_position', 'fen': fen,
                                                                          'actor': self.name}))
+
         self.thread_active = True
         self.event_thread = threading.Thread(target=self._event_worker_thread)
         self.event_thread.setDaemon(True)
@@ -79,9 +72,6 @@ class Protocol(ParserCallback, CalibrationCallback):
             self._read_config()
         except Exception as e:
             logger.debug(f'No valid default configuration, starting board-scan: {e}')
-
-        self.parser = CertaboBoardMessageParser(self, self.low_gain)
-
         while True:
             if not self.device_in_config:
                 self._search_board()
@@ -89,7 +79,7 @@ class Protocol(ParserCallback, CalibrationCallback):
             if self.config is None or self.trans is None:
                 logger.error('Cannot connect.')
                 if self.config is None:
-                    self.config = {'low_gain': self.low_gain}
+                    self.config = {}
                     self.write_configuration()
                 self.error_condition = True
             else:
@@ -101,9 +91,11 @@ class Protocol(ParserCallback, CalibrationCallback):
             time.sleep(3)
 
     def _read_config(self):
-        with open('certabo_config.json', 'r') as f:
+        with open('chessnut_config.json', 'r') as f:
             self.config = json.load(f)
-            self.low_gain = self.config.get('low_gain', True)
+            if 'btle_iface' not in self.config:
+                self.config['btle_iface'] = 0
+                self.write_configuration()
             if 'address' in self.config:
                 logger.debug(f"Checking default configuration for board at {self.config['address']}")
                 trans = self._open_transport()
@@ -112,37 +104,43 @@ class Protocol(ParserCallback, CalibrationCallback):
                     self.trans = trans
 
     def _search_board(self):
-        tr = Transport(self.trque)
-        logger.debug('created obj')
-        if tr.is_init():
-            logger.debug('Transport loaded.')
-            address = tr.search_board()
-            if address is not None:
-                logger.debug(f'Found board at address {address}')
-                self.config = {'address': address, 'low_gain': self.low_gain}
-                self.trans = tr
-                self.write_configuration()
-        else:
-            logger.warning('Failed to initialize')
+        try:
+            tr = Transport(self.trque, READ_CHARACTERISTIC, WRITE_CHARACTERISTIC)
+            logger.debug('created obj')
+            if tr.is_init():
+                logger.debug('Transport loaded.')
+                if self.config is not None:
+                    btle = self.config['btle_iface']
+                else:
+                    btle = 0
+                address = tr.search_board('Chessnut', btle)
+                if address is not None:
+                    logger.debug(f'Found board at address {address}')
+                    self.config = {'address': address}
+                    self.trans = tr
+                    self.write_configuration()
+            else:
+                logger.warning('Bluetooth failed to initialize')
+        except Exception as e:
+            logger.warning(f'Internal error, import of Bluetooth failed: {e}')
 
     def _connect(self):
         address = self.config['address']
         logger.debug(f'Valid board available at {address}')
-        logger.debug(f'Connecting to Certabo at {address}')
+        logger.debug(f'Connecting to Chessnut at {address}')
         self.connected = self.trans.open_mt(address)
         if self.connected:
-            logger.info(f'Connected to Certabo at {address}')
-            self.led_control = CertaboLedControl(self.trans)
-            self.sentio = Sentio(self, self.led_control)
-            self.error_condition = False
+            logger.info(f'Connected to Chessnut at {address}')
         else:
             self.trans.quit()
-            logger.error(f'Connection to Certabo at {address} FAILED.')
+            logger.error(f'Connection to Chessnut at {address} FAILED.')
+            self.config = {}
+            self.write_configuration()
             self.error_condition = True
 
     def quit(self):
         """
-        Quit Certabo connection.
+        Quit Chessnut connection.
         Try to terminate transport threads gracefully.
         """
         if self.trans is not None:
@@ -151,7 +149,7 @@ class Protocol(ParserCallback, CalibrationCallback):
 
     def position_initialized(self):
         """
-        Check, if a board position has been received and the Certabo board is online.
+        Check, if a board position has been received and the Chessnut board is online.
 
         :return: True, if board position has been received
         """
@@ -163,19 +161,26 @@ class Protocol(ParserCallback, CalibrationCallback):
         return False
 
     def write_configuration(self):
+        """
+        Write the configuration for Bluetooth LE to 'chessnut_config.json'.
+
+        :return: True on success, False on error
+        """
+        if 'btle_iface' not in self.config:
+            self.config['btle_iface'] = 0
         try:
-            with open('certabo_config.json', 'w') as f:
+            with open('chessnut_config.json', 'w') as f:
                 json.dump(self.config, f, indent=4)
                 return True
         except Exception as e:
-            logger.error(f'Failed to save default configuration {self.config} to certabo_config.json: {e}')
+            logger.error(f'Failed to save default configuration {self.config} to chessnut_config.json: {e}')
         return False
 
     def _event_worker_thread(self):
         """
         The event worker thread is automatically started during __init__.
         """
-        logger.debug('Certabo worker thread started.')
+        logger.debug('Chessnut worker thread started.')
         while self.thread_active:
             if not self.trque.empty():
                 msg = self.trque.get()
@@ -189,7 +194,8 @@ class Protocol(ParserCallback, CalibrationCallback):
                     else:
                         state = toks
                         emsg = ''
-                    logger.info(f'Agent state of {self.name} changed to {state}, {emsg}')
+                    logger.info(
+                        f'Agent state of {self.name} changed to {state}, {emsg}')
                     if state == 'offline':
                         self.error_condition = True
                     else:
@@ -197,34 +203,18 @@ class Protocol(ParserCallback, CalibrationCallback):
                     self.appque.put({'cmd': 'agent_state', 'state': state, 'message': emsg})
                     continue
 
-                if not self.calibrated and self.piece_recognition:
-                    self.calibrator.calibrate(msg)
-                else:
-                    self.parser.parse(msg)
+                self.parser.parse(msg)
             else:
                 time.sleep(0.01)
 
-    def has_piece_recognition(self, piece_recognition: bool):
-        self.piece_recognition = True
-        if piece_recognition:
-            self.calibrate()
-
-    def occupied_squares(self, board: List[int]):
-        if self.sentio is not None:
-            self.sentio.occupied_squares(board)
-
     def board_update(self, short_fen: str):
-        if not self.initial_position_received and short_fen == 'rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR':
-            self.initial_position_received = True
-        if self.initial_position_received:
-            # only forward the fen if the initial position has been received at least once
-            # to prevent additional queens on the board from changing settings
-            self.debouncer.update(short_fen)
+        self.debouncer.update(short_fen)
         with self.board_mutex:
             self.last_fen = short_fen
 
-    def request_promotion_dialog(self, move: str):
-        self.appque.put({'cmd': 'request_promotion_dialog', 'move': move, 'actor': self.name})
+    def battery(self, percent: int, status: Battery):
+        msg = f'{status.name} {percent}'
+        self.appque.put({'cmd': 'battery', 'message': msg})
 
     def reversed(self, value: bool):
         self.brd_reversed = value
@@ -233,47 +223,28 @@ class Protocol(ParserCallback, CalibrationCallback):
         """
         Set LEDs according to `position`.
 
-        :param pos: `position` array, field != 0 indicates a LED that should be on
+        :param pos: `position` array, field != 0 indicates a led that should be on
         """
-        if self.connected and self.led_control is not None:
-            cmd = command.set_leds(pos, self.brd_reversed)
-            self.led_control.write_led_command(cmd)
+        if self.connected:
+            cmd = command.set_led(pos, self.brd_reversed)
+            self.trans.write_mt(cmd)
 
     def set_led_off(self):
-        if self.connected and self.led_control is not None:
-            self.led_control.write_led_command(command.set_leds_off())
-
-    def uci_move(self, move: str):
         if self.connected:
-            self.sentio.uci_move(move)
+            self.trans.write_mt(command.set_led_off())
 
-    def promotion_done(self, uci_move: str):
+    def request_battery_status(self):
         if self.connected:
-            self.sentio.promotion_done(uci_move)
+            self.trans.write_mt(command.request_battery_status())
 
-    def calibrate(self):
+    def realtime_mode(self):
         if self.connected:
-            self.calibrated = False
-            if self.led_control is not None:
-                self.led_control.write_led_command(command.set_leds_calibrate())
-
-    def calibration_complete(self, stones: typing.Dict[CertaboPiece, typing.Optional[str]]):
-        logger.info('Certabo calibration complete')
-        self.parser.update_stones(stones)
-        self.calibrated = True
-        self.set_led_off()
-
-    def calibration_complete_square(self, square: int):
-        # can be used to clear LED of square during calibration, not implemented
-        pass
-
-    def calibration_error(self):
-        pass
+            self.trans.write_mt(command.request_realtime_mode())
 
     def _open_transport(self):
-        tr = Transport(self.trque)
+        tr = Transport(self.trque, READ_CHARACTERISTIC, WRITE_CHARACTERISTIC)
         if tr.is_init():
             return tr
         else:
-            logger.warning('USB transport failed to initialize')
+            logger.warning('Bluetooth transport failed to initialize')
         return None
