+++ eboard/ichessone/board.py
@@ -16,18 +16,19 @@ import logging
 from threading import Thread
 import queue
 
-from eboard import EBoard
+from eboard.eboard import EBoard
 from utilities import DisplayMsg
 from dgt.api import Message, Dgt
 from dgt.util import ClockIcons
 
-from certabo.certabo_agent import CertaboAgent
+from eboard.ichessone.ichessone_agent import IChessOneAgent
+from eboard.ichessone.parser import Battery
 
 
 logger = logging.getLogger(__name__)
 
 
-class CertaboBoard(EBoard):
+class IChessOneBoard(EBoard):
 
     def __init__(self):
         self.agent = None
@@ -35,16 +36,15 @@ class CertaboBoard(EBoard):
 
     def light_squares_on_revelation(self, uci_move: str):
         logger.debug('turn LEDs on - move: %s', uci_move)
-        dpos = [[0 for _ in range(8)] for _ in range(8)]
+        dpos = [[0 for x in range(8)] for y in range(8)]
         dpos[int(uci_move[1]) - 1][ord(uci_move[0]) - ord('a')] = 1  # from
         dpos[int(uci_move[3]) - 1][ord(uci_move[2]) - ord('a')] = 1  # to
         if self.agent is not None:
-            self.agent.uci_move(uci_move)
             self.agent.set_led(dpos)
 
     def light_square_on_revelation(self, square: str):
         logger.debug('turn on LEDs - square: %s', square)
-        dpos = [[0 for _ in range(8)] for _ in range(8)]
+        dpos = [[0 for x in range(8)] for y in range(8)]
         dpos[int(square[1]) - 1][ord(square[0]) - ord('a')] = 1
         if self.agent is not None:
             self.agent.set_led(dpos)
@@ -64,40 +64,61 @@ class CertaboBoard(EBoard):
             except queue.Empty:
                 pass
             bwait = waitchars[wait_counter]
-            text = self._display_text('no Certabo e-Board' + bwait, 'Certabo' + bwait, 'Certabo' + bwait, bwait)
+            text = self._display_text('no IChessOne e-Board' + bwait, 'IChessOne' + bwait, 'IChess1' + bwait, bwait)
             DisplayMsg.show(Message.DGT_NO_EBOARD_ERROR(text=text))
             wait_counter = (wait_counter + 1) % len(waitchars)
             time.sleep(1.0)
 
         if result['state'] != 'offline':
             logger.info('incoming_board ready')
-
         self._process_after_connection()
 
     def _process_after_connection(self):
+        last_battery_request = time.time()
         while True:
             if self.agent is not None:
                 try:
                     result = self.appque.get(block=False)
                     if 'cmd' in result and result['cmd'] == 'agent_state' and 'state' in result and 'message' in result:
-                        if result['state'] == 'offline':
-                            text = self._display_text(result['message'], result['message'], 'no/', 'Board')
-                        else:
-                            text = Dgt.DISPLAY_TIME(force=True, wait=True, devs={'ser', 'i2c', 'web'})
-                        DisplayMsg.show(Message.DGT_NO_EBOARD_ERROR(text=text))
+                        self._process_board_state(result)
                     elif 'cmd' in result and result['cmd'] == 'raw_board_position' and 'fen' in result:
-                        fen = result['fen'].split(' ')[0]
-                        DisplayMsg.show(Message.DGT_FEN(fen=fen, raw=True))
-                    elif 'cmd' in result and result['cmd'] == 'request_promotion_dialog' and 'move' in result:
-                        DisplayMsg.show(Message.PROMOTION_DIALOG(move=result['move']))
+                        self._process_board_position(result)
+                    elif 'cmd' in result and result['cmd'] == 'battery' and 'message' in result:
+                        self._process_battery_state(result)
+
                 except queue.Empty:
                     pass
+                current_time = time.time()
+                if current_time - last_battery_request > 30:  # request battery state every 30 seconds
+                    last_battery_request = current_time
+                    self.agent.request_battery_status()
 
             time.sleep(0.1)
 
+    def _process_board_state(self, result):
+        if result['state'] == 'offline':
+            text = self._display_text(result['message'], result['message'], 'no Board', 'no brd')
+        else:
+            self.agent.request_board_updates()
+            text = Dgt.DISPLAY_TIME(force=True, wait=True, devs={'ser', 'i2c', 'web'})
+        DisplayMsg.show(Message.DGT_NO_EBOARD_ERROR(text=text))
+
+    def _process_board_position(self, result):
+        fen = result['fen'].split(' ')[0]
+        DisplayMsg.show(Message.DGT_FEN(fen=fen, raw=True))
+
+    def _process_battery_state(self, result):
+        if Battery.LOW.name in result['message'] or Battery.EXHAUSTED.name in result['message']:
+            text = self._display_text('Battery ' + result['message'], 'Batt.' + result['message'], 'batt low', 'batlow')
+            DisplayMsg.show(Message.DGT_NO_EBOARD_ERROR(text=text))
+        else:
+            battery_str = result['message'].split()[1]
+            if battery_str.isnumeric():
+                DisplayMsg.show(Message.BATTERY(percent=int(battery_str)))
+
     def _connect(self):
         logger.info('connecting to board')
-        self.agent = CertaboAgent(self.appque)
+        self.agent = IChessOneAgent(self.appque)
 
     def set_text_rp(self, text: bytes, beep: int):
         return True
@@ -127,4 +148,4 @@ class CertaboBoard(EBoard):
         pass
 
     def promotion_done(self, uci_move: str):
-        self.agent.promotion_done(uci_move)
+        pass
