+++ eboard/ichessone/parser.py
@@ -12,14 +12,9 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 from typing import List
-from enum import Enum
 
-
-class Battery(Enum):
-    CHARGING = 0
-    DISCHARGING = 1
-    LOW = 2
-    EXHAUSTED = 3
+from eboard.eboard import to_short_fen, to_battery, get_upper_4_bits, get_lower_4_bits, check_reversed
+from eboard.eboard import Battery
 
 
 class ParserCallback(object):
@@ -56,11 +51,11 @@ class Parser(object):
                         if len(board) > 0 and None not in board:
                             if self.last_board != board:
                                 self.last_board = board
-                                board = self._check_reversed(board)
-                                self.callback.board_update(self._to_short_fen(board))
+                                board, self.reversed = check_reversed(board, self.reversed, self.callback)
+                                self.callback.board_update(to_short_fen(board))
                             i = new_pos
                     elif (i + 3) < len(data) and data[i] == 0x3d and data[i + 1] == 0x62:  # battery
-                        self.callback.battery(*self._to_battery(data[i + 3], data[i + 2]))
+                        self.callback.battery(*to_battery(data[i + 3], data[i + 2]))
                         i += 3
                 i += 1
         else:
@@ -79,32 +74,22 @@ class Parser(object):
     def _add_to_buffer(self, arr):
         self.buffer += arr
 
-    def _to_battery(self, percent, status):
-        battery = Battery.DISCHARGING
-        if status == 1:
-            battery = Battery.CHARGING
-        value = max(0, min(100, percent))
-        if value < 5:
-            battery = Battery.EXHAUSTED
-        elif value < 10:
-            battery = Battery.LOW
-        return value, battery
-
     def _to_board(self, data):
         if len(data) >= 32:
             board = [None] * 64
             i = 0
             for row in range(7, -1, -1):
                 for col in range(4):
-                    board[i] = self._to_stone(self._get_upper_4_bits(data[row * 4 + col]))
-                    board[i + 1] = self._to_stone(self._get_lower_4_bits(data[row * 4 + col]))
+                    board[i] = self._to_stone(get_upper_4_bits(data[row * 4 + col]))
+                    board[i + 1] = self._to_stone(get_lower_4_bits(data[row * 4 + col]))
                     i += 2
             i += 1
             return board
         else:
             return []
 
-    def _to_stone(self, value):
+    @staticmethod
+    def _to_stone(value):
         translation = {0: ' ',
                        0x01: 'P', 0x04: 'R', 0x02: 'N', 0x03: 'B', 0x05: 'Q', 0x06: 'K',
                        0x07: 'p', 0x0a: 'r', 0x08: 'n', 0x09: 'b', 0x0b: 'q', 0x0c: 'k'}
@@ -112,54 +97,3 @@ class Parser(object):
             return translation[value]
         else:
             return None
-
-    def _to_short_fen(self, board):
-        fen = ''
-        for row_index, row in enumerate(range(7, -1, -1)):
-            blanks = 0
-            for col_index, col in enumerate(range(8)):
-                if board[row * 8 + col] == ' ':
-                    blanks += 1
-                    if col_index == 7 and blanks > 0:
-                        fen += str(blanks)
-                else:
-                    if blanks > 0:
-                        fen += str(blanks)
-                    blanks = 0
-                    fen += board[row * 8 + col]
-            if row_index != 7:
-                fen += '/'
-        return fen
-
-    def _check_reversed(self, brd):
-        board = brd.copy()
-        w_count_lower_half, b_count_lower_half = self._piece_count(board, range(32))
-        w_count_upper_half, b_count_upper_half = self._piece_count(board, range(32, 64))
-        if self.reversed and w_count_lower_half > 10 and b_count_upper_half > 10:
-            self.reversed = False
-            self.callback.reversed(self.reversed)
-        elif not self.reversed and w_count_upper_half > 10 and b_count_lower_half > 10:
-            self.reversed = True
-            self.callback.reversed(self.reversed)
-        if self.reversed:
-            board.reverse()
-        return board
-
-    def _piece_count(self, board, board_half):
-        w_count = 0
-        b_count = 0
-        for i in board_half:
-            if board[i] != ' ':
-                if board[i] < 'Z':
-                    w_count += 1
-                else:
-                    b_count += 1
-        return w_count, b_count
-
-    @staticmethod
-    def _get_upper_4_bits(b):
-        return (b & 0xf0) >> 4
-
-    @staticmethod
-    def _get_lower_4_bits(b):
-        return b & 0x0f
